+++ main.ts
@@ -4,27 +4,36 @@ import { ConfirmModal } from './src/ui/components/ConfirmModal';
 import { SettingsTab } from './src/ui/settings/SettingsTab';
 import { PluginSettings, DEFAULT_SETTINGS } from './src/types/settings';
 import { LLMService } from './src/services/llm-service';
-import { FileService } from './src/services/file-service';
+import { FoodLogManager } from './src/services/food-log-manager';
+import * as MealOps from './src/services/meal/manager';
 import { applyEmojiPreferences } from './src/utils/apply-emoji-preferences';
 import { migrateIfNeeded } from './src/services/meal/migrate-if-needed';
 
 export default class NutritionTrackerPlugin extends Plugin {
   settings: PluginSettings;
   llmService: LLMService;
-  fileService: FileService;
+  foodLogManager: FoodLogManager;
   private editButtonHandler: (event: Event) => void;
   private deleteButtonHandler: (event: Event) => void;
   private ctaButtonHandler: (event: Event) => void;
   private mealSyncTimeouts: Map<string, number> = new Map();
   private isDeleteInProgress: boolean = false;
-  private currentModal: FoodInputModal | null = null; 
+  private currentModal: FoodInputModal | null = null;
+
+  private get mealDeps(): MealOps.MealDeps {
+    return {
+      vault: this.app.vault,
+      app: this.app,
+      settings: this.settings
+    };
+  }
 
   async onload() {
     await this.loadSettings();
 
     // Initialize services
     this.llmService = new LLMService(this.settings);
-    this.fileService = new FileService(this.app, this.app.vault, this.settings);
+    this.foodLogManager = new FoodLogManager(this.app.vault, this.settings);
 
     // Run meal migration
     await migrateIfNeeded(this.app.vault, this.settings);
@@ -73,7 +82,7 @@ export default class NutritionTrackerPlugin extends Plugin {
     this.registerEvent(
       this.app.vault.on('rename', async (file, oldPath) => {
         if (file instanceof TFile) {
-          await this.fileService.handleFileRename(oldPath, file.path);
+          await MealOps.handleFileRename(this.mealDeps, oldPath, file.path);
         }
       })
     );
@@ -118,12 +127,13 @@ export default class NutritionTrackerPlugin extends Plugin {
 
   private createAndOpenModal(setupFn?: (modal: FoodInputModal) => void) {
     this.ensureModalClosed();
-    
+
     this.currentModal = new FoodInputModal(
-      this.app, 
-      this.settings, 
-      this.llmService, 
-      this.fileService,
+      this.app,
+      this.app.vault,
+      this.settings,
+      this.llmService,
+      this.foodLogManager,
       () => {
         this.currentModal = null;
       }
@@ -173,10 +183,10 @@ export default class NutritionTrackerPlugin extends Plugin {
     
     try {
       if (context === 'meal') {
-        await this.fileService.deleteMealItem(entryToDelete);
+        await MealOps.deleteMealItem(this.mealDeps, entryToDelete);
         new Notice(`🍽️ Deleted from meal: ${food} (${quantity})`);
       } else {
-        await this.fileService.deleteFoodLogItem(entryToDelete);
+        await this.foodLogManager.deleteFoodLogItem(entryToDelete);
         new Notice(`📝 Deleted from food log: ${food} (${quantity})`);
       }
       
@@ -307,21 +317,21 @@ export default class NutritionTrackerPlugin extends Plugin {
   private setupMealNoteSyncHandler() {
     this.registerEvent(
       this.app.vault.on('modify', (file) => {
-        const isMealNote = this.fileService.isMealNote(file);
-        
+        const isMealNote = MealOps.isMealNote(this.mealDeps, file);
+
         if (isMealNote && file instanceof TFile) {
           const filePath = file.path;
-          
+
           if (this.mealSyncTimeouts.has(filePath)) {
             window.clearTimeout(this.mealSyncTimeouts.get(filePath));
           }
-          
+
           const timeout = window.setTimeout(() => {
-            void this.fileService.syncMealNoteToJSON(file).then(() => {
+            void MealOps.syncMealNoteToJSON(this.mealDeps, file).then(() => {
               this.mealSyncTimeouts.delete(filePath);
             });
           }, 1000);
-          
+
           this.mealSyncTimeouts.set(filePath, timeout);
         }
       })
@@ -341,17 +351,17 @@ export default class NutritionTrackerPlugin extends Plugin {
       this.llmService = new LLMService(this.settings);
     }
 
-    if (this.fileService) {
-      this.fileService = new FileService(this.app, this.app.vault, this.settings);
+    if (this.foodLogManager) {
+      this.foodLogManager = new FoodLogManager(this.app.vault, this.settings);
     }
   }
 
   async saveSettings() {
     await this.saveData(this.settings);
-    
+
     this.llmService = new LLMService(this.settings);
-    this.fileService = new FileService(this.app, this.app.vault, this.settings);
-    
+    this.foodLogManager = new FoodLogManager(this.app.vault, this.settings);
+
     applyEmojiPreferences(this.settings.appearance);
   }
