+++ src/ui/components/FoodInputModal/utils/meal-management.ts
@@ -1,15 +1,29 @@
+import { Vault, App } from 'obsidian';
 import { Meal } from '../../../../types/nutrition';
-import { FileService } from '../../../../services/file-service';
+import { PluginSettings } from '../../../../types/settings';
+import * as MealOps from '../../../../services/meal/manager';
 
 export class MealManager {
   private availableMeals: Meal[] = [];
   private selectedMeals: Meal[] = [];
 
-  constructor(private fileService: FileService) {}
+  constructor(
+    private vault: Vault,
+    private app: App,
+    private settings: PluginSettings
+  ) {}
+
+  private get mealDeps(): MealOps.MealDeps {
+    return {
+      vault: this.vault,
+      app: this.app,
+      settings: this.settings
+    };
+  }
 
   async loadMeals(): Promise<void> {
     try {
-      this.availableMeals = await this.fileService.getMeals();
+      this.availableMeals = await MealOps.getMeals(this.mealDeps);
     } catch (error) {
       console.error('Error loading meals:', error);
       this.availableMeals = [];
@@ -17,7 +31,7 @@ export class MealManager {
   }
 
   async addMeal(mealId: string): Promise<void> {
-    const meal = await this.fileService.getMealById(mealId);
+    const meal = await MealOps.getMealById(this.mealDeps, mealId);
     if (meal && !this.selectedMeals.find(m => m.id === meal.id)) {
       this.selectedMeals.push(meal);
     }
